+++ topobank/publication/urls.py
@@ -1,4 +1,5 @@
-from django.urls import path
+from django.contrib.auth.decorators import login_required
+from django.urls import re_path, path
 
 from rest_framework.routers import DefaultRouter
 
@@ -20,5 +21,20 @@ urlpatterns += [
         '<str:short_url>/download/',
         view=views.download,
         name='go-download'
-    )
+    ),
+    re_path(
+        r'publish/(?P<pk>\d+)/$',
+        view=login_required(views.SurfacePublishView.as_view()),
+        name='surface-publish'
+    ),
+    re_path(
+        r'publish/(?P<pk>\d+)/publication-rate-too-high/$',
+        view=login_required(views.PublicationRateTooHighView.as_view()),
+        name='surface-publication-rate-too-high'
+    ),
+    re_path(
+        r'publish/(?P<pk>\d+)/publication-error/$',
+        view=login_required(views.PublicationErrorView.as_view()),
+        name='surface-publication-error'
+    ),
 ]
